adalasso {adalasso}	R Documentation

Master function to use the adaptive lasso for shrinkage and selection in a generalized linear model

Description

Master function to use the adaptive lasso for shrinkage and selection in a generalized linear model

Usage

adalasso(formula, data, family, weights, max.iter = 20, tol = 1e-25, s = NULL)

Arguments

formula
The model formula

data
Model data frame

family
To which exponential-family distribution does the GLM belong?

weights
Observation weights

max.iter
How many times to iterate the adaptive-lasso cycle

tol
Tolerance for conversion

s
Values of the lasso tuning parameter at which to fit models

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (formula, data, family, weights, max.iter = 20, tol = 1e-25, 
    s = NULL) 
{
    result = list()
    class(result) = "adalasso"
    result[["formula"]] = as.formula(formula, env = data)
    model.data = data
    na.rows = (which(is.na(model.data)) - 1)%%dim(model.data)[1] + 
        1
    if (length(na.rows) > 0) 
        model.data = model.data[-na.rows, ]
    response.name = rownames(attr(terms(formula, data = data), 
        "factors"))[1]
    predictor.names = attr(terms(formula, data = data), "term.labels")
    response.col = which(names(model.data) == response.name)
    y = as.matrix(model.data[, response.col])
    x = as.matrix(model.data[, -response.col])
    result[["lasso"]] = lasso_step(y = y, x = x, family = family, 
        weights = weights, s = s)
    iter = 1
    change = tol + 1
    lambda.former = result[["lasso"]][["lambda"]]
    result[["lambda"]] = c(lambda.former)
    while (iter <= max.iter && change > tol) {
        cat(paste("Iteration: ", iter, "\n", sep = ""))
        f = as.formula(paste(paste(response.name, "~", sep = ""), 
            paste(result[["lasso"]][["vars"]], collapse = "+"), 
            sep = ""))
        result[["adapt"]] = adapt = adaptive_step(formula = f, 
            data = model.data, family = family, weights = weights)
        result[["lasso"]] = lasso_step(y = y, x = as.matrix(model.data[, 
            -response.col]), family = family, weights = weights, 
            adaptive.object = adapt, s = s)
        result[["lambda"]] = c(result[["lambda"]], result[["lasso"]][["lambda"]])
        change = abs(lambda.former - tail(result[["lambda"]], 
            1))
        lambda.former = tail(result[["lambda"]], 1)
        iter = iter + 1
        cat(paste("Change in lambda: ", change, "\n", sep = ""))
    }
    result[["iter"]] = iter - 1
    return(result)
  }

[Package adalasso version 0.1 Index]