adalasso-package {adalasso}	R Documentation

Use the adaptive lasso for shrinkage and selection in a generalized linear model

Description

Iterate between glmnet (for shrinkage/selection) and glm (for adaptive penalty weights) to get an adaptive lasso estimate of coefficients in a GLM where p>n.

Details

Package: 	 adalasso

Type: 	 Package

Version: 	 0.1

Date: 	 2012-07-03

License: 	 GPL-3

~~ An overview of how to use the package, including the most important functions ~~

Author(s)

Wesley Brooks

Maintainer: Wesley Brooks <wrbrooks@usgs.gov> ~~ The author and/or maintainer of the package ~~

References

~~ Literature or other references for background information ~~

See Also

~~ Optional links to other man pages, e.g. ~~ ~~ <pkg> ~~

Examples

~~ simple examples of the most important functions ~~

[Package adalasso version 0.1 Index]