adalasso_initial_step {adalasso}	R Documentation

The first glm step uses each variable as a predictor, one at a time.

Description

The first glm step uses each variable as a predictor, one at a time.

Usage

adalasso_initial_step(formula, data, family, weights, ...)

Arguments

formula

data

family

weights

...

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (formula, data, family, weights, ...) 
{
    result <- list()
    response.name = rownames(attr(terms(formula, data = data), 
        "factors"))[1]
    predictor.names = attr(terms(formula, data = data), "term.labels")
    na.rows = (which(is.na(data)) - 1)%%dim(data)[1] + 1
    if (length(na.rows) > 0) 
        data = data[-na.rows, ]
    result[["data"]] = data
    result[["model"]] = model = glm(formula = formula, data = data, 
        family = family, weights = weights, ...)
    coefs = as.list(coef(model))
    coefs[["(Intercept)"]] = NULL
    adaweight = vector()
    for (name in names(data)[-which(names(data) == response.name)]) {
        if (name %in% predictor.names) {
            adaweight = c(adaweight, 1/coefs[[name]])
        }
        else {
            adaweight = c(adaweight, 1)
        }
    }
    result[["coefs"]] = coefs
    result[["adaweight"]] = adaweight
    return(result)
  }

[Package adalasso version 0.1 Index]