adalasso_step {adalasso}	R Documentation

The glmnet step (shrinks and selects model coefficients) for adaptive lasso shrinkage and selection in a generalized linear model

Description

The glmnet step (shrinks and selects model coefficients) for adaptive lasso shrinkage and selection in a generalized linear model

Usage

adalasso_step(y, x, family, weights, adaptive.object = NULL, s = NULL, ...)

Arguments

y

x

family

weights

adaptive.object

s

...

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (y, x, family, weights, adaptive.object = NULL, s = NULL, 
    ...) 
{
    result = list()
    m <- ncol(x)
    n <- nrow(x)
    one <- rep(1, n)
    result[["meanx"]] = (meanx <- drop(one %*% x)/n)
    x.centered <- scale(x, meanx, FALSE)
    normx <- sqrt(drop(one %*% (x.centered^2)))
    names(normx) <- NULL
    xs = x.centered
    for (k in 1:dim(x.centered)[2]) {
        if (normx[k] != 0) {
            xs[, k] = xs[, k]/normx[k]
        }
        else {
            xs[, k] = rep(0, dim(xs)[1])
            normx[k] = Inf
        }
    }
    if (is.null(adaptive.object)) {
        result[["coef.scale"]] = coef.scale = 1/normx
    }
    else {
        ada.weight = adaptive.object[["adaweight"]]
        for (k in 1:dim(x.centered)[2]) {
            if (!is.na(ada.weight[k])) {
                xs[, k] = xs[, k] * ada.weight[k]
            }
            else {
                xs[, k] = rep(0, dim(xs)[1])
                ada.weight[k] = 0
            }
        }
        result[["coef.scale"]] = coef.scale = ada.weight/normx
    }
    result[["model"]] = model = glmnet(x = xs, y = y, family = family, 
        weights = weights, lambda = s, ...)
    result[["cv.model"]] = cv.model = cv.glmnet(y = y, x = xs, 
        nfolds = dim(x)[1], family = family, weights = weights, 
        lambda = s, ...)
    result[["lambda"]] = lambda = cv.model$lambda.min
    nonzero = as.vector(predict(model, type = "nonzero", s = lambda))
    print(nonzero)
    coefs = coef(model, s = lambda)
    result[["coef"]] = as.list(coefs)[nonzero + 1]
    names(result[["coef"]]) = rownames(coefs)[nonzero + 1]
    result[["intercept"]] = intercept = coefs[1]
    result[["vars"]] = names(result[["coef"]])
    return(result)
  }

[Package adalasso version 0.1 Index]